requirements.txt
fastapi==0.98.0
opencv_python==4.7.0.72
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.114
uvicorn==0.22.0
python-multipart